Split source code.
