+++ Sources/Builder/Failure.swift
@@ -1,0 +1,15 @@
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+// Created by Sam Deane, 02/03/2018.
+// All code (c) 2018 - present day, Elegant Chaos Limited.
+// For licensing terms, see http://elegantchaos.com/license/liberal/.
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+/**
+ Errors generated by failures at various stages of the build process.
+ */
+
+enum Failure : Error {
+    case failed(output : String?, error : String?)
+    case decodingFailed
+}
+
